--- src/layouts/AuthLayout/index.tsx
@@ -1,11 +1,0 @@
-import { ReactNode } from "react"
-
-const AuthLayout = ({ children }: { children: ReactNode }) => {
-  return (
-    <div>
-      <main>{children}</main>
-    </div>
-  )
-}
-
-export default AuthLayout
