+++ src/pages/Auth/Login/Login.tsx
@@ -17,7 +17,7 @@ const Login = () => {
     return (
 
         <div className="flex justify-center items-center h-full">
-            <div className="w-full max-w-[456px] h-auto p-6 sm:p-8 md:p-10 lg:p-12 rounded-2xl shadow absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 bg-white">
+            <div className="w-full max-w-[456px] h-auto p-6 sm:p-8 md:p-10 lg:p-12 rounded-2xl shadow mt-[100px] bg-white">
                 <div>
                     <h1 className="font-bold text-DSM text-center">Đăng nhập</h1>
                     <p className="my-5 text-center text-gray-400">
